+++ android/app/src/main/java/com/voidroot/bikeos/data/backup/BackupManager.kt
@@ -236,6 +236,7 @@ class BackupManager @Inject constructor(
                 put("avgCadenceRpm", r.avgCadenceRpm)
                 put("maxCadenceRpm", r.maxCadenceRpm)
                 put("rideMode", r.rideMode)
+                put("avgAccelJerkG", r.avgAccelJerkG)
             })
         }
     }.toString()